refactor pm quality summary lookup routes
