+++ src/app/router_registry.py
@@ -73,6 +73,9 @@ from app.routers.dpm_command_center_pm_quality_review_actions import (
 from app.routers.dpm_command_center_pm_quality_summary_invocations import (
     router as dpm_command_center_pm_quality_summary_invocations_router,
 )
+from app.routers.dpm_command_center_pm_quality_summary_lookup import (
+    router as dpm_command_center_pm_quality_summary_lookup_router,
+)
 from app.routers.dpm_command_center_portfolio_memory import (
     router as dpm_command_center_portfolio_memory_router,
 )
@@ -243,6 +246,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
     app.include_router(dpm_command_center_pm_quality_summary_invocations_router)
+    app.include_router(dpm_command_center_pm_quality_summary_lookup_router)
     app.include_router(dpm_command_center_portfolio_memory_router)
     app.include_router(dpm_wave_campaign_definitions_router)
     app.include_router(dpm_wave_campaign_discovery_router)
